+++ service/src/main/java/bio/terra/profile/service/profile/flight/create/CreateProfileFlight.java
@@ -9,7 +9,6 @@ import bio.terra.profile.service.gcp.GcpService;
 import bio.terra.profile.service.iam.SamService;
 import bio.terra.profile.service.job.JobMapKeys;
 import bio.terra.profile.service.policy.TpsApiDispatch;
-import bio.terra.profile.service.profile.flight.common.VerifyUserBillingAccountAccessStep;
 import bio.terra.profile.service.profile.model.BillingProfile;
 import bio.terra.profile.service.profile.model.ProfileDescription;
 import bio.terra.stairway.Flight;
