+++ service/src/test/java/bio/terra/profile/service/profile/flight/create/CreateProfileFlightUnitTest.java
@@ -12,7 +12,6 @@ import bio.terra.profile.service.azure.AzureService;
 import bio.terra.profile.service.crl.GcpCrlService;
 import bio.terra.profile.service.iam.SamService;
 import bio.terra.profile.service.job.JobMapKeys;
-import bio.terra.profile.service.profile.flight.common.VerifyUserBillingAccountAccessStep;
 import bio.terra.stairway.FlightMap;
 import java.util.List;
 import java.util.UUID;
